requirements.txt
streamlit
groq
python-dotenv
requests
beautifulsoup4
langchain-text-splitters
